# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCP Content Processor",
    "version": "1.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/process": {
      "post": {
        "summary": "Process Content",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text":     { "type": "string" },
                  "platform": { "type": "string" },
                  "pillar":   { "type": "string" },
                  "name":     { "type": "string" },
                  "brand":    { "type": "string", "default": "Ethical AI Insider" },
                  "context":  { "type": "string" }
                },
                "required": ["text", "platform", "pillar"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Processed content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "formatted":       { "type": "string" },
                    "safe":            { "type": "string" },
                    "headlines": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "platform":        { "type": "string" },
                    "pillar":          { "type": "string" },
                    "brand":           { "type": "string" },
                    "brandCompliance": { "type": "string" },
                    "name":            { "type": "string" },
                    "context":         { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens": {
      "post": {
        "summary": "Token Count",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string" },
                  "model": { "type": "string" }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokens": { "type": "integer" },
                    "model": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}